fix(neovim): texlab integration

+++ .config/nvim/after/lsp/texlab.lua
@@ -9,7 +9,7 @@
 local venv_bin, path, venv = vim.fs.root(0, ".venv/bin/python")
 if venv_bin then
   path = ("%s:%s"):format(venv_bin, vim.env.PATH)
-  venv = venv_bin
+  venv = vim.fs.dirname(venv_bin)
 end
 
 local log_win = -1
@@ -57,32 +57,70 @@ return {
     }
   },
   on_init = function(client)
-    -- define for easy log access
+    local augroup = vim.api.nvim_create_augroup(("%s_%d"):format(client.name, client.id), {})
+
+    local function on_attach(bufnr, aux_dir)
+      vim.api.nvim_buf_create_user_command(
+        bufnr,
+        "TexlabLog",
+        function() TexlabLog(vim.fs.joinpath(client.config.root_dir or ".", aux_dir)) end,
+        {desc = "Show content of log files in a floating window."}
+      )
+
+      -- define some keybindings for convenient access to some lsp commands
+      vim.keymap.set("n", "<localleader>bn", '<cmd>LspTexlabBuild<cr>', {silent = true, buffer = bufnr})
+      vim.keymap.set("n", "<localleader>fs", '<cmd>LspTexlabForward<cr>', {silent = true, buffer = bufnr})
+      vim.keymap.set("n", "<localleader>sl", '<cmd>TexlabLog<cr>', {silent = true, buffer = bufnr})
+      vim.keymap.set("n", "<localleader>ce", '<cmd>LspTexlabChangeEnvironment<cr>', {silent = true, buffer = bufnr})
+      vim.keymap.set("n", "<localleader>cl", '<cmd>LspTexlabCleanAuxiliary<cr>', {silent = true, buffer = bufnr})
+      vim.keymap.set("n", "<localleader>cn", '<cmd>LspTexlabCancelBuild<cr>', {silent = true, buffer = bufnr})
+    end
+
+    local function on_dir_report(obj)
+      -- find actual aux_dir following latexmk approach: https://github.com/latex-lsp/texlab/pull/968
+      local dirs = {obj.stdout:match("Normalized aux dir, out dir, out2 dir:%s-'(.-)', '(.-)', '(.-)'")}
+      -- even when using --dir-report-only latexmk still creates these
+      -- directories which might not be what we want
+      for _, path in ipairs(dirs) do
+        if vim.fs.abspath(path) ~= vim.fs.abspath(client.config.root_dir or ".") then
+          os.remove(path)
+        end
+      end
+
+      -- help texlab find the compiled pdf for forward search
+      client.config.settings.texlab.build.pdfDirectory = dirs[3]
+      client.notify("workspace/didChangeConfiguration",
+        {settings = client.config.settings}
+      )
+
+      vim.api.nvim_create_autocmd("LspAttach", {
+        group = augroup,
+        callback = function(args)
+          if args.data.client_id == client.id then
+            on_attach(args.buf, dirs[1])
+          end
+        end
+      })
+
+      -- manually run for already attached buffers
+      for buf, _ in ipairs(client.attached_buffers) do
+        on_attach(buf, dirs[1])
+      end
+
+      vim.api.nvim_create_autocmd("LspDetach", {
+        group = augroup,
+        callback = function(args)
+          if args.data.client_id == client.id then
+            vim.api.nvim_clear_autocmds{group=augroup}
+          end
+        end,
+      })
+    end
+
     vim.system(
       {client.config.settings.texlab.build.executable, "-dir-report-only"},
       {text = true},
-      vim.schedule_wrap(function(obj)
-        -- find actual aux_dir following latexmk approach: https://github.com/latex-lsp/texlab/pull/968
-        local aux_dir, out_dir = obj.stdout:match("Normalized aux dir, out dir, out2 dir:%s-'(.-)', '(.-)'")
-        -- even when using --dir-report-only latexmk still creates these
-        -- directories which might not be what we want
-        for _, path in ipairs{aux_dir, out_dir} do
-          os.remove(path)
-        end
-        vim.api.nvim_create_user_command(
-          "TexlabLog",
-          function() TexlabLog(vim.fs.joinpath(client.config.root_dir or ".", aux_dir)) end,
-          {desc = "Show content of log files in a floating window."}
-        )
-      end)
+      vim.schedule_wrap(on_dir_report)
     )
-
-    -- define some keybindings for convenient access to some lsp commands
-    vim.keymap.set("n", "<localleader>bn", '<cmd>LspTexlabBuild<cr>', {silent = true, buffer = true})
-    vim.keymap.set("n", "<localleader>fs", '<cmd>LspTexlabForward<cr>', {silent = true, buffer = true})
-    vim.keymap.set("n", "<localleader>sl", '<cmd>TexlabLog<cr>', {silent = true, buffer = true})
-    vim.keymap.set("n", "<localleader>ce", '<cmd>LspTexlabChangeEnvironment<cr>', {silent = true, buffer = true})
-    vim.keymap.set("n", "<localleader>cl", '<cmd>LspTexlabCleanAuxiliary<cr>', {silent = true, buffer = true})
-    vim.keymap.set("n", "<localleader>cn", '<cmd>LspTexlabCancelBuild<cr>', {silent = true, buffer = true})
   end
 }
